Create a parameter defining the performed risk stratified analysis

Create a parameter defining the performed risk stratified analysis

Usage

createAnalysisSettings(
  analysisId = NULL,
  description = "",
  databaseName,
  treatmentCohortId,
  comparatorCohortId,
  outcomeIds,
  analysisMatrix = diag(length(outcomeIds)),
  mapTreatments,
  mapOutcomes,
  negativeControlOutcomes = c(),
  balanceThreads = 1,
  negativeControlThreads = 1,
  verbosity = NULL,
  saveDirectory = NULL
)

Arguments

analysisId

The analysis ID.

description

Text describing the analysis.

databaseName

The name of the database.

treatmentCohortId

The cohort definition id of the treatment cohort in the cohortTable.

comparatorCohortId

The cohort definition id of the comparator cohort in the cohortTable.

outcomeIds

The cohort definition ids of the outcome cohorts in the outcomeTable.

analysisMatrix

Boolean matrix defining the outcomes to be assessed (rows) within risk strata (columns). The order in columns should match the the order of outcomeIds. Default is the diagonal matrix, which leads to the risk stratified assessment of only the outcome for which the risk strata were defined.

mapTreatments

Dataframe containing 2 columns: "exposure_id" with the id numbers of the treatment and comparator cohorts and "exposure_name" the cohort names.

mapOutcomes

Dataframe containing 2 columns: "outcome_id" with the cohort names of the outcomes of interest and "outcome_name" with their names.

negativeControlOutcomes

The outcome Ids to be used as negative controls

balanceThreads

The number of threads to be used for the estimation of covariate balance

negativeControlThreads

The number of threads to be used for running the negative control analyses

verbosity

Sets the level of the verbosity. If the log level is at or higher in priority than the logger threshold, a message will print. The levels are:

DEBUGHighest verbosity showing all debug statements

TRACEShowing information about start and end of steps

INFOShow informative information (Default)

WARNShow warning messages

ERRORShow error messages

FATALBe silent except for fatal errors

saveDirectory

The directory name where the results of the analyses will be stored.

Value

An analysisSettings object providing the identification information of the analysis.